Auth moved to controller

+++ todo-backend/app/Http/Controllers/TodoController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreTodos;
 use App\Http\Requests\UpdateTodos;
 use App\Services\TodoService;
 use App\Todo as Todo;
+use Illuminate\Support\Facades\Auth;
 
 class TodoController extends Controller
 {
@@ -25,7 +26,7 @@ class TodoController extends Controller
 
     public function index()
     {
-        return $this->todoService->getAll();
+        return $this->todoService->getAll(Auth::user());
     }
 
     public function show(Todo $todo){
@@ -35,16 +36,20 @@ class TodoController extends Controller
 
     public function store(StoreTodos $request)
     {
-        return $this->todoService->addNew($request);
+        $todo = array_merge(['user_id'=>Auth::id()], $request->validated());
+        return $this->todoService->addNew($todo);
     }
 
     public function update(UpdateTodos $request, Todo $todo)
     {
-        return $this->todoService->update($request, $todo);
+        $this->authorize('update', $todo);
+        $valid = $request->validated();
+        return $this->todoService->update($valid, $todo);
     }
 
     public function destroy(Todo $todo)
     {
+        $this->authorize('delete', $todo);
         $this->todoService->delete($todo);
     }
 }
